feat(copytree): auto-select context format based on focused agent
- Create shared format selection helper in src/lib/copyTreeFormat.ts
- Auto-select XML for Claude/Codex, Markdown for Gemini based on focused terminal
- Update worktree.copyTree action to determine format from ActionContext
- Include format in user feedback message (e.g., "Copied 42 files as XML")
- Refactor useContextInjection to use shared helper for consistency
