+++ src/services/actions/definitions/worktreeActions.ts
@@ -4,6 +4,8 @@ import type { ActionContext, ActionId } from "@shared/types/actions";
 import { copyTreeClient, githubClient, systemClient, worktreeClient } from "@/clients";
 import { useWorktreeDataStore } from "@/store/worktreeDataStore";
 import { useWorktreeSelectionStore } from "@/store/worktreeStore";
+import { useTerminalStore } from "@/store/terminalStore";
+import { getFormatForTerminal } from "@/lib/copyTreeFormat";
 
 export function registerWorktreeActions(actions: ActionRegistry, callbacks: ActionCallbacks): void {
   actions.set("worktree.refresh", () => ({
@@ -373,15 +375,20 @@ export function registerWorktreeActions(actions: ActionRegistry, callbacks: Acti
       format: z.enum(["xml", "json", "markdown", "tree", "ndjson"]).optional(),
     }),
     run: async (args: unknown, ctx: ActionContext) => {
-      const { worktreeId, format } = args as {
+      const { worktreeId, format: explicitFormat } = args as {
         worktreeId?: string;
         format?: "xml" | "json" | "markdown" | "tree" | "ndjson";
       };
       const targetWorktreeId = worktreeId ?? ctx.activeWorktreeId;
       if (!targetWorktreeId) return null;
 
+      const terminal = ctx.focusedTerminalId
+        ? useTerminalStore.getState().terminals.find((t) => t.id === ctx.focusedTerminalId)
+        : undefined;
+      const format = explicitFormat ?? getFormatForTerminal(terminal);
+
       const result = await copyTreeClient.generateAndCopyFile(targetWorktreeId, {
-        format: format ?? "xml",
+        format,
       });
 
       if (result.error) {
@@ -392,6 +399,7 @@ export function registerWorktreeActions(actions: ActionRegistry, callbacks: Acti
         worktreeId: targetWorktreeId,
         fileCount: result.fileCount,
         stats: result.stats ?? null,
+        format,
       };
     },
   }));
